Interactive grid using static variable
The grid can now be clicked to activate/deactivate cells in the grid. I
used a static variable in a class called GameGrid to hold the
information, because I was experiencing a similar thing to Riz where the
variable in the controller would reset everytime. It saves the changes
as long as the browser is on the page, but restarting it clears the
grid.